added a selected link component
